tests/test_scrapper.py
# tests/test_scraper.py
"""
Unit tests for Twitter scraper
"""

import pytest
from src.collectors.twitter_scraper import TwitterScraper
import yaml


@pytest.fixture
def config():
    """Load test configuration"""
    with open('config.yaml', 'r') as f:
        return yaml.safe_load(f)


@pytest.fixture
def scraper(config):
    """Create scraper instance"""
    return TwitterScraper(config)


def test_scraper_initialization(scraper):
    """Test scraper initializes correctly"""
    assert scraper is not None
    assert scraper.driver is not None
    assert scraper.tweets_collected == []


def test_parse_tweet_element():
    """Test tweet parsing logic"""
    # This would need a mock HTML element
    pass


def test_human_scroll(scraper):
    """Test scroll behavior"""
    # Mock test - would need browser instance
    assert hasattr(scraper, '_human_scroll')
