requirements.txt
# Web Scraping
selenium==4.27.1
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Data Processing
pandas==2.2.3
pyarrow==18.1.0
numpy==2.2.1

# NLP & ML
scikit-learn==1.6.0
nltk==3.9.1
emoji==2.14.0

# Visualization
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
tqdm==4.67.1
colorlog==6.9.0

# Performance
joblib==1.4.2

# Testing
pytest==8.3.4
pytest-cov==6.0.0